Added velocity measuring function in ESP32Servo360 lib
Function getSpeed() has been changed to get a real instantaneous angular velocity in (rotation per minute); previous getSpeed() was renamed to getStatedSpeed(), returns the speed set by default.

+++ src/FW/gpio.h
@@ -28,7 +28,7 @@ int init_LServo(void);
 void rotate_LServo(float speed);
 int get_LServoAngle(void);
 void set_LServoOffset(int offsetAngle);
-int getLServoSpeed(void);
+float getLServoSpeed(void);
 
 #ifdef __cplusplus
